pyproject.toml
[tool.poetry]
name = "smart-pharmacy-backend"
version = "0.1.0"
description = ""
authors = ["MIT"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
pandas = "^2.2.2"
openpyxl = "^3.1.2"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
openai = "^1.30.5"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
